# Point2D implements ISpanFormattable