--- a/Projects/Server/Geometry/Point2D.cs
+++ b/Projects/Server/Geometry/Point2D.cs
@@ -14,7 +14,6 @@
  *************************************************************************/
 
 using System;
-using Server.Buffers;
 
 namespace Server;